do not run clippy for cargo clippy -- -W help #6122
